+++ changelog.md
@@ -10,6 +10,7 @@
 ### Improvements
 * #1981 Added grouping for cart rules.
 * #1407 Added menu to set order and payment status for orders checked in backend grid.
+* #1778: Back button closes offcanvas menu/cart.
 
 
 ## Smartstore 4.0.1